Рефакторит отрисовку дней и точек маршрута
